requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.0.1
django-cors-headers==4.3.1
django-filter==23.3
